requirements.txt
pytest
allure-pytest
opencv-python
pyzbar
pdfminer.six
pdf2image
pydantic
jsonschema